Add mock for lps-oppfolgingsplan-mottak
Fikser mocking ved kjøring lokalt.
Ref: #1238
