+++ .env.template
@@ -36,6 +36,8 @@ SYFOMOTEBEHOV_AAD_APP_CLIENT_ID=71e608a6-fad4-42f6-b8cf-0130c82fa78a
 SYFOMOTEBEHOV_HOST=syfomotebehov.dev.intern.nav.no
 SYFOOPPFOLGINGSPLANSERVICE_AAD_APP_CLIENT_ID=7d8c5d52-348f-48c2-b42f-f59cf6970fd3
 SYFOOPPFOLGINGSPLANSERVICE_HOST=syfooppfolgingsplanservice.dev.intern.nav.no
+LPS_OPPFOLGINGSPLAN_MOTTAK_AAD_APP_CLIENT_ID=dev-gcp.team-esyfo.lps-oppfolgingsplan-mottak
+LPS_OPPFOLGINGSPLAN_MOTTAK_HOST=lps-oppfolgingsplan-mottak.dev.intern.nav.no
 SYFOPERSON_AAD_APP_CLIENT_ID=dev-gcp.teamsykefravr.syfoperson
 SYFOPERSON_HOST=syfoperson.dev.intern.nav.no
 SYFOSMREGISTER_AAD_APP_CLIENT_ID=543ceb1e-eb69-4089-9458-bdec61160afa
